requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
pyttsx3==2.98
groq==0.12.0
mysql-connector-python==9.1.0
